move ut for SetNamespaceCommand to V2CommandsPrivilegesSuite
